+++ src/app/Components/ArtworkRail/ArtworkRailCard.tsx
@@ -7,9 +7,9 @@ import { CreateSavedSearchModal } from "app/Components/Artist/ArtistArtworks/Cre
 import { saleMessageOrBidInfo as defaultSaleMessageOrBidInfo } from "app/Components/ArtworkGrids/ArtworkGridItem"
 import { useExtraLargeWidth } from "app/Components/ArtworkRail/useExtraLargeWidth"
 import { ContextMenuArtwork } from "app/Components/ContextMenu/ContextMenuArtwork"
+import { useArtworkContextMenu } from "app/Components/ContextMenu/useArtworkContextMenu"
 import OpaqueImageView from "app/Components/OpaqueImageView/OpaqueImageView"
 import { getUrgencyTag } from "app/utils/getUrgencyTag"
-import { useArtworkContextMenu } from "app/utils/hooks/useArtworkContextMenu"
 import { useSaveArtwork } from "app/utils/mutations/useSaveArtwork"
 import { Schema } from "app/utils/track"
 import { sizeToFit } from "app/utils/useSizeToFit"
